requirements.txt
boto>=2.5.2
django>=1.4.1
django-admin-sso>=0.1.1
django-cache-url>=0.3
django-cors-headers==0.05
django-crispy-forms>=1.1.4
django-debug-toolbar>=0.9.4
django-extensions==0.9
django-never-cache-post>=0.1.1
django-sslify==0.2
django-storages==1.1.4
djangorestframework==2.1.16
dj-database-url>=0.2
gunicorn>=0.14.5
psycopg2>=2.4.5
pylibmc>=1.2.3
raven>=2.0.0
south>=0.7.6
